Initial commit
The bot currently cycles through a list of names, inserting them into the search bar and selecting enter. If the corresponding PNG file can be found on screen, the bot prints that it has found it. Otherwise it prints that it has not found it. 
Features to include will be price reading functionality and the ability to report via Discord. For example, if asset X drops below Y price, send a message to me.